pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-notion-reading-inbox"
version = "0.1.0"
description = "Local-first Telegram to Notion reading inbox tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27.0",
  "pydantic>=2.7.0",
  "PyYAML>=6.0.1",
  "beautifulsoup4>=4.12.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
